Create hover state and grid configuration button
Created mouseover event listener for individual divs that changes
background color to mimic hover effect.  Also reworked code into
various functions and created a configuration button.  That way
when you change the aspects of the grid, the old grid is removed and
a new grid is created.  I also allowed the separate configuration of
the number of containers and divs.